moved PlaybackSetupData to a separate file
