+++ src/framework/mpe/CMakeLists.txt
@@ -26,6 +26,7 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/soundid.h
     ${CMAKE_CURRENT_LIST_DIR}/mpetypes.h
     ${CMAKE_CURRENT_LIST_DIR}/events.h
+    ${CMAKE_CURRENT_LIST_DIR}/playbacksetupdata.h
     ${CMAKE_CURRENT_LIST_DIR}/iarticulationprofilesrepository.h
 
     ${CMAKE_CURRENT_LIST_DIR}/view/articulationpatternsegmentitem.cpp
